BeastshopWebParent/BeastshopBackEnd/src/main/java/com/beastshop/admin/category/CategoryRepository.java
package com.beastshop.admin.category;

import java.util.List;

import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.PagingAndSortingRepository;

import com.beastshop.common.entity.Category;

public interface CategoryRepository extends PagingAndSortingRepository<Category, Integer> {

	@Query("SELECT c FROM Category c WHERE c.parent.id is NULL")
	public List<Category> findRootCategories();
	
	
	//method by spring data jpa
	public Category findByName(String name);
	
	public Category findByAlias(String alias);
}
